board detail 작성자 이름 표시 구현 및 boardListVO 구현중

+++ khist/src/main/java/com/kh/khist/dto/AttachDto.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class AttachDto {
 	private int attachNo;
 	private String attachName;
-	private int attachSize;
+	private long attachSize;
 	private String attachType;
 	private Date attachDate;
